+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Revision history for pqueue
 
+## 1.5.0
+
+  * Add pattern synonyms to work with `MinQueue` and `MinPQueue`.
+    ([#92](http://github.com/lspitzner/pqueue/pull/92))
+
+  * Make the `Data` instances respect the queue invariants. Make the
+    `Constr`s match the pattern synonyms. Make the `Data` instance for
+    `MinPQueue` work "incrementally", like the one for `MinQueue`.
+    ([#92](http://github.com/lspitzner/pqueue/pull/92))
+
 ## 1.4.3.0 -- 2022-10-30
 
   * Add instances for [indexed-traversable](https://hackage.haskell.org/package/indexed-traversable).